// frontend/src/client/models/Macros.ts
// tslint:disable
/**
 * Macro Macro API
 * An API that provides access to recipe and ingredient information, both user-generated and sourced from the USDA
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Macros
 */
export interface Macros {
    /**
     * 
     * @type {number}
     * @memberof Macros
     */
    fat: number;
    /**
     * 
     * @type {number}
     * @memberof Macros
     */
    carbs: number;
    /**
     * 
     * @type {number}
     * @memberof Macros
     */
    protein: number;
    /**
     * 
     * @type {number}
     * @memberof Macros
     */
    calories: number;
}

export function MacrosFromJSON(json: any): Macros {
    return MacrosFromJSONTyped(json, false);
}

export function MacrosFromJSONTyped(json: any, ignoreDiscriminator: boolean): Macros {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'fat': json['fat'],
        'carbs': json['carbs'],
        'protein': json['protein'],
        'calories': json['calories'],
    };
}

export function MacrosToJSON(value?: Macros): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'fat': value.fat,
        'carbs': value.carbs,
        'protein': value.protein,
        'calories': value.calories,
    };
}